backend/requirements.txt
fastapi
uvicorn
ollama
apscheduler
supabase
python-dotenv
requests
pydantic
pandas
python-multipart
openpyxl
